update game objects

+++ src/Beryll/Core/EventHandler.cpp
@@ -90,7 +90,7 @@ namespace Beryll
                                                  glm::vec2(event.tfinger.x * Window::getInstance()->getScreenWidth(), event.tfinger.y * Window::getInstance()->getScreenHeight()),
                                                   false,
                                                   true,
-                                                  event.tfinger.fingerId});
+                                                  static_cast<long>(event.tfinger.fingerId)});
                     break;
 
                 case SDL_FINGERUP:
